+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update Terraform AWS provider to 3.x [#329](https://github.com/PublicMapping/districtbuilder/pull/329)
 - Extend Jenkins pipeline to support production releases [#325](https://github.com/PublicMapping/districtbuilder/pull/325)
 - Add support for Rollbar [#338](https://github.com/PublicMapping/districtbuilder/pull/338)
+- Provide math for scaling task memory with additional states [#309](https://github.com/PublicMapping/districtbuilder/pull/309)